pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trdr"
version = "0.0.4"
authors = [
    { name = "developer_jeb", email = "developer_jeb@outlook.com" },
]
requires-python = ">=3.12"
description = "Algorithmic trading library"
dependencies = [
    "yfinance",
    "aiohttp==3.11.11",
    "opentelemetry-api==1.30.0",
    "opentelemetry-sdk==1.30.0",
    "pydantic==2.10.6",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"
